Conflicts:
	mysql-test/r/show_check.result
	mysql-test/suite/innodb/r/innodb.result
	mysql-test/t/show_check.test
	sql/dd/impl/properties_impl.h
	sql/item_func.cc
	storage/innobase/handler/ha_innodb.cc

+++ storage/innobase/os/os0file.cc
@@ -3906,7 +3906,7 @@ SyncFileIO::execute(Slot* slot)
 
 	}
 
-	return(ret ? slot->n_bytes : -1);
+	return(ret ? static_cast<ssize_t>(slot->n_bytes) : -1);
 }
 
 /** Check if the file system supports sparse files.
